# api/openapi.yaml
openapi: 3.0.3
info:
  title: guided-ssh API
  description: |
    REST API of the guided-ssh server: certificate issuance (human + CI),
    host enrollment, grant management, and the read/audit endpoints of the
    web UI. This specification is the single source of truth; the
    Angular API client is generated from it (`make web-api`).

    Authentication: OIDC ID token as bearer token. Roles of the admin API
    (phase 8): `admin` (everything), `auditor` (read/export audit + read),
    `readonly` (read resources); roles are mapped to IdP groups and are
    inclusive of one another (admin ⊃ auditor ⊃ readonly).
  version: "1.0"
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /
...
security:
  - bearerAuth: []
paths:
...
  /v1/admin/enroll-tokens:
    post:
      tags: [rollout]
      summary: Mint enrollment token for the one-command install (role admin)
      description: |
        The plaintext token is returned exactly once (never logged, never
        included in the audit payload). Gated like the other rollout
        endpoints.
      operationId: createEnrollToken
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EnrollTokenRequest" }
      responses:
        "201":
          description: Token, expiry, and ready-to-use install command
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnrollTokenResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/RolloutUnavailable" }
...
  /v1/admin/service-accounts/{id}:
    patch:
      tags: [directory]
      summary: Activate/deactivate service account — kill switch (role admin)
      operationId: updateServiceAccount
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [active]
              properties:
                active: { type: boolean }
      responses:
        "200":
          description: Updated service account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServiceAccount" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/admin/certificates:
    get:
      tags: [directory]
      summary: Most recently issued certificates (role readonly)
      operationId: listCertificates
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 500 } }
      responses:
        "200":
          description: Certificate metadata, newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Certificate" }
  /v1/admin/audit:
    get:
      tags: [audit]
      summary: Filter audit events (role auditor)
      description: |
        Filterable by event type, actor, time range, and full text (`q`
        matches actor and payload — covers host and pipeline filters).
      operationId: listAudit
      parameters:
        - { name: event_type, in: query, schema: { type: string } }
        - { name: actor, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string }, description: Substring match over actor and payload }
        - { name: since, in: query, schema: { type: string, format: date-time } }
        - { name: until, in: query, schema: { type: string, format: date-time } }
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 500 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Filtered events with total count
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuditList" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /v1/admin/audit/export:
    get:
      tags: [audit]
      summary: Audit export as CSV or JSON download (role auditor)
      operationId: exportAudit
      parameters:
        - { name: format, in: query, schema: { type: string, enum: [json, csv], default: json } }
        - { name: event_type, in: query, schema: { type: string } }
        - { name: actor, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string } }
        - { name: since, in: query, schema: { type: string, format: date-time } }
        - { name: until, in: query, schema: { type: string, format: date-time } }
      responses:
        "200":
          description: Download (max. 100000 rows)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AuditEvent" }
            text/csv:
              schema: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: OIDC-ID-Token
  responses:
    BadRequest:
      description: Invalid request
      content:
        text/plain: { schema: { type: string } }
    Unauthorized:
      description: Token missing or invalid
      content:
        text/plain: { schema: { type: string } }
    Forbidden:
      description: Insufficient role/permission
      content:
        text/plain: { schema: { type: string } }
    NotFound:
      description: Not found
      content:
        text/plain: { schema: { type: string } }
    RolloutUnavailable:
      description: Host rollout not configured (missing conditions in `missing`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/RolloutUnavailable" }
  schemas:
...
    RolloutUnavailable:
      type: object
      required: [error, missing]
      properties:
        error: { type: string }
        missing: { type: array, items: { type: string } }
    EnrollTokenRequest:
      type: object
      properties:
        hostname: { type: string, description: empty ⇒ token not bound to a hostname }
        tags:
          type: object
          additionalProperties: { type: string }
        ttl_seconds: { type: integer, format: int64, default: 3600, minimum: 60, maximum: 86400 }
        session_audit:
          type: boolean
          description: Only sets --session-audit in install_command; not stored on the token
    EnrollTokenResponse:
      type: object
      required: [token, expires_at, install_command]
      properties:
        token: { type: string, description: Plaintext, delivered exactly once }
        expires_at: { type: string, format: date-time }
        install_command: { type: string }
...
    ServiceAccount:
      type: object
      required: [id, name, kind, issuer, claim_matcher, active, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        kind: { type: string }
        issuer: { type: string }
        claim_matcher:
          type: object
          additionalProperties: { type: string }
        active: { type: boolean, description: false acts as a kill switch per project }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Certificate:
      type: object
      required: [id, serial, key_id, cert_type, principals, valid_after, valid_before, created_at]
      properties:
        id: { type: string, format: uuid }
        serial: { type: integer, format: int64 }
        key_id: { type: string }
        cert_type: { type: string, enum: [user, host] }
        principals: { type: array, items: { type: string } }
        valid_after: { type: string, format: date-time }
        valid_before: { type: string, format: date-time }
        issuer_context:
          type: object
          description: Context of issuance (SSO session, CI pipeline claims)
        created_at: { type: string, format: date-time }
    AuditEvent:
      type: object
      required: [id, occurred_at, event_type, actor, payload]
      properties:
        id: { type: integer, format: int64 }
        occurred_at: { type: string, format: date-time }
        event_type:
          type: string
          description: e.g. cert.issued, grant.created/updated/deleted, host.enrolled, user.deactivated, service_account.updated
        actor: { type: string }
        payload:
          type: object
          description: Event details as JSON
    AuditList:
      type: object
      required: [events, total]
      properties:
        events:
          type: array
          items: { $ref: "#/components/schemas/AuditEvent" }
        total:
          type: integer
          format: int64
          description: Total count for the filter (for pagination)